% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asVol.R
\name{asVol}
\alias{asVol}
\alias{asVol.data.table}
\alias{asVol.flowLoad}
\alias{asVol.rainLoad}
\title{Add convert flow or rainfall data to a volume}
\usage{
asVol(x, area)

\method{asVol}{data.table}(x, area = NULL)

\method{asVol}{flowLoad}(x, area = NULL)

\method{asVol}{rainLoad}(x, area = NULL)
}
\arguments{
\item{x}{Data set of 'flowLoad' or 'rainLoad' class}

\item{area_km}{Area of coverage for rain gauges}
}
\value{
Calculated volume of flow or rainfall
}
\description{
Add convert flow or rainfall data to a volume
}
\examples{
# oundle <- loadWISKI(files[4])
# oundle
# asVol.rainLoad(oundle, area = 40)
# oundle
#
#
# link <- 'C:/Users/jpayne05/Desktop/Buildwas_15min_Flow.csv'
# Buildwas <- loadWISKI(link)
# Buildwas
# asVol.flowLoad(Buildwas)
# Buildwas
}
