% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumSkipNA.R
\name{cumsumNA}
\alias{cumsumNA}
\title{cumsumNA}
\usage{
cumsumNA(x)

cumsumNA(x)
}
\arguments{
\item{x}{Raw rain gauge data or merged dataset}

\item{...}{Additional parameters as required}
}
\value{
A cumulative sum of rain gauge data
}
\description{
Provides a cumulative sum of rain gauge data while excluding NA
values.
}
\examples{
cumsumNA(rain)
}
