% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyAgg.R
\name{dailyAgg}
\alias{dailyAgg}
\alias{dailyAgg.data.table}
\title{dailyAgg}
\usage{
dailyAgg(x, method = "mean", ...)

\method{dailyAgg}{data.table}(x, method = "mean", ...)
}
\arguments{
\item{x}{Data generated in the riskyData package}

\item{method}{'mean', 'median', 'max', 'min', or 'sum'}

\item{...}{Other variables as required}
}
\value{
Data are aggregated to a daily resolution
}
\description{
Aggregates sub daily time series into a daily resolution,
aggregations are carried out to the calendar day not a rolling time period.
}
\examples{
dailyAgg(Buildwas)
}
