% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAgg.R
\name{dataAgg}
\alias{dataAgg}
\title{Data aggregation for \code{HydroImport} objects}
\usage{
dataAgg(x = NULL, type = NULL, method = NULL)
}
\arguments{
\item{x}{A data.table that sits within an R6 \code{HydroImport} object}

\item{type}{Set the aggregation level to either hourly', 'daily', 'monthly',
'annual', or 'hydroYear'}

\item{method}{Available aggregation methods include 'min', 'max', 'mean',
'median', and 'sum'}
}
\value{
Aggregated dataset determined from the input \code{type} and \code{method}. Data are placed within a \code{HydroAggs} type container
}
\description{
This acts as a wrapper function for all the aggregation
functions that sit within riskyData
}
\examples{
# Load bewdley dataset
data(bewdley)

# Calculate mean hourly flows
bewdley$dataAgg(type = "hourly", method = "mean")

# Calculate median daily flows
bewdley$dataAgg(type = "daily", method = "median")

# Calculate min
bewdley$dataAgg(type = "monthly", method = "min")

# Calculate annual (calendar year) max
bewdley$dataAgg(type = "annual", method = "max")

# Calculate hydroYear max
# Hydrological years need to be calculated for this. The `hydroYearDay()` function can be piped into the operation.
bewdley$hydroYearDay()$dataAgg(type = "hydroYear", method = "max")

# Calculate total hourly flow
# Flow has to be converted to volume. The `asVol()` function can be piped into the operation.
bewdley$asVol()$dataAgg(type = "hourly", method = "sum")
}
\seealso{
\link{hourlyAgg}

\link{dailyAgg}

\link{monthlyAgg}

\link{annualAgg}

\link{hydroYearAgg}
}
