% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayStats.R
\name{dayStats}
\alias{dayStats}
\alias{dayStats.data.table}
\title{dayStats}
\usage{
dayStats(
  x,
  methods = c("mean", "median", "min", "max"),
  percentiles = c(5, 25, 75, 95),
  hydroDay = TRUE,
  plot = TRUE,
  ...
)

\method{dayStats}{data.table}(
  x = NULL,
  methods = c("mean", "median", "min", "max"),
  percentiles = c(5, 25, 75, 95),
  hydroDay = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{x}{Data generated in the riskyData package}

\item{methods}{'mean', 'median', 'max', 'min', or 'sum'}

\item{percentiles}{Set to 5, 25, 75, and 95}

\item{plot}{Set to TRUE, will produce a plot of the statistics}

\item{...}{Other variables as required}

\item{hyrdroDay}{Set to TRUE, use hydrological day}
}
\value{
Data are aggregated to a daily resolution
}
\description{
Daily statistics for imported data
}
\examples{
# dayStats(Buildwas)
}
