% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceed.R
\name{exceed}
\alias{exceed}
\title{Find the amount of events over a given threshold and supply the amount of
time steps to which they were in excess.}
\usage{
exceed(
  dateTime = NULL,
  value = NULL,
  threshold = NULL,
  gapWidth = 0,
  timeStep = 900
)
}
\arguments{
\item{dateTime}{The date and time series}

\item{value}{The observed data}

\item{threshold}{The threshold at which to test}

\item{gapWidth}{This allows you to ignore a set number of time steps beneath
a threshold between two or more events}

\item{timeStep}{The time step used applied in the observed data in seconds}
}
\value{
A table of events with start and end times
}
\description{
Find the amount of events over a given threshold and supply the amount of
time steps to which they were in excess.
}
\examples{
data(bewdley)
events <- exceed(bewdley$data$dateTime,
                 bewdley$data$value,
                 threshold = 200,
                 gapWidth = 20,
                 timeStep = 900)
events
}
