% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportToPDM.R
\name{exportToPDM}
\alias{exportToPDM}
\title{Export data in a format suitable for PDM for PCs}
\usage{
exportToPDM(x = NULL, path = NULL, type = "csv", name = "PDM_Input")
}
\arguments{
\item{x}{File, currently only supports the class \code{catchAvg} for rainfall}

\item{path}{Optional path for file export. Defaults to working directory.}

\item{type}{Export type that supports csv, XML, and data.table (for internal use)}

\item{name}{Catchment name used in XML exports}
}
\value{
csv, XML or data.tables in the format suitable for PDM for PCs
}
\description{
Export data in a format suitable for PDM for PCs
}
\examples{
## Import some data
loadAPI(ID = c("590310", "197036", "U06"),
        measure = 'rainfall',
        period = 900,
        type = 'total',
        datapoints = 'range',
        from = '2023-10-01 09:00',
        to = '2024-10-01 08:45')
## Carry out a catchment average
rainfall <- catchAvg(rainfall_197036, rainfall_590310, rainfall_U06, areas = c(45, 25, 15))
exportToPDM(x = rainfall, type = "data.table")
}
