% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaks.R
\name{findPeaks}
\alias{findPeaks}
\title{Find peaks in hydrometric data}
\usage{
findPeaks(x, levels = 100, from = NULL, to = NULL, gaps = 96)
}
\arguments{
\item{x}{R6 hydroImport object}

\item{levels}{The number of thresholds to apply against the time seires.
Defaults to 100}

\item{from}{Set the location from where the thresholds should start.
Defaults to the bottom 10\% value.}

\item{to}{Set the location from where the thresholds should end
Defaults to the maximum observed value.}

\item{gaps}{Used to pad out the run length encoding, prevents wobbly data
showing multiple exceedances. Set to 96.}
}
\value{
A data.table containing dateTime and values of observed peaks
}
\description{
The \code{findPeaks} function detects peaks in hydrometric time
series using run length encoding against a subset of threshlds.
}
\examples{
data(bewdley)
findPeaks(bewdley)
}
