% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDMCArho.R
\name{getDMCArho}
\alias{getDMCArho}
\title{Function to perform the DMCA analysis}
\usage{
getDMCArho(
  Rdata,
  Qdata,
  Lmin = 1,
  Lmax = 400,
  dL = 2,
  fracLperDt = 1,
  minWindowsPerL = 5
)
}
\arguments{
\item{Rdata}{Rdata is cumulative rainfall data. A vector, no time data.}

\item{Qdata}{Qdata is cumulative flow or level data. A vector, no time data.}

\item{Lmin}{mallest time window. Must be an ODD integer.}

\item{Lmax}{Largest time window.}

\item{dL}{Increment between windows. Must be an EVEN integer.}

\item{fracLperDt}{Set the time increment as a fraction of window size.}

\item{minWindowsPerL}{Minimum number of windows in the data. Replaces Lmax if needed.}
}
\value{
Minima representing the estimated Tp
}
\description{
Function to perform the DMCA analysis
}
\examples{
## Do not run
}
\author{
Oliver Dyer
}
