% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGaugeSAAR.R
\name{getGaugeSAAR}
\alias{getGaugeSAAR}
\title{Claculate the observed SAAR values from rain gauge data}
\usage{
getGaugeSAAR(gaugeIDs)
}
\arguments{
\item{gaugeIDs}{A vector of rain gauge WISKI IDs that can be fed into
\code{riskyData::loadAPI()}}
}
\value{
A list containing the annual rainfalls, summary statistics and plots
}
\description{
Claculate the observed SAAR values from rain gauge data
}
\examples{
rainGaugeIDs <- c('015347','021228', '013045', 'NE034')
getGaugeSAAR(rainGaugeIDs)
}
