% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdeAPI.R
\name{hdeAPI}
\alias{hdeAPI}
\title{Download hydrometric data from the Hydrology Data Explorer}
\usage{
hdeAPI(
  ID = NULL,
  measure = NULL,
  period = NULL,
  type = NULL,
  datapoints = "standard",
  from = NULL,
  to = NULL,
  rtExt = FALSE,
  meta = TRUE,
  rtLookup = FALSE
)
}
\arguments{
\item{ID}{Use to specify a particular WISKI ID or for when downloading all
stations 'wiski' or 'nrfa' to filter results. Additionally to access the
real time APIs use "flood" for all critical sites and "tidal" for the
category A tide gauges.}

\item{measure}{Use this when exporting observations to select the available
parameter. Generally 'flow', 'level', or 'groundwater'}

\item{period}{This is so you can select the time steps available, generally
900 (15 min) or 86400 (daily)}

\item{type}{Selects the data type, instantaneous, min, max etc.}

\item{datapoints}{Allows you to decide where in the time series you are
pulling the data from}

\item{from}{First time step of choice}

\item{to}{Last time step of choice}

\item{rtExt}{Set to NULL, if TRUE hydrological data series are extended with
the real time API}

\item{meta}{Set as TRUE, exports metadata for gauge with the data request}

\item{rtLookup}{Set as FALSE, this uses an inbuilt dataset, if set to TRUE
this will update the table using the most recent data. However, this can
dramatically increase the run time.}
}
\value{
r6 HydroImport container
}
\description{
Used in the \code{loadAPI()} function
}
\examples{
#hdeAPI(ID = '2001', measure = 'flow', period = 900)
}
