% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroImportFactory.R
\name{HydroImport}
\alias{HydroImport}
\alias{HydroImportFactory}
\title{R6 object for imported data via the EA API}
\value{
A class containing raw data, metadata and methods
}
\description{
This is the base class for all data imported via the API. A HydroImport contains public raw data and private metadata.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Imported data via the API tool. Uses data.table.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HydroImport-new}{\code{HydroImportFactory$new()}}
\item \href{#method-HydroImport-print}{\code{HydroImportFactory$print()}}
\item \href{#method-HydroImport-methods}{\code{HydroImportFactory$methods()}}
\item \href{#method-HydroImport-summary}{\code{HydroImportFactory$summary()}}
\item \href{#method-HydroImport-asVol}{\code{HydroImportFactory$asVol()}}
\item \href{#method-HydroImport-hydroYearDay}{\code{HydroImportFactory$hydroYearDay()}}
\item \href{#method-HydroImport-rmVol}{\code{HydroImportFactory$rmVol()}}
\item \href{#method-HydroImport-rmHY}{\code{HydroImportFactory$rmHY()}}
\item \href{#method-HydroImport-rmHYD}{\code{HydroImportFactory$rmHYD()}}
\item \href{#method-HydroImport-coords}{\code{HydroImportFactory$coords()}}
\item \href{#method-HydroImport-nrfa}{\code{HydroImportFactory$nrfa()}}
\item \href{#method-HydroImport-dataAgg}{\code{HydroImportFactory$dataAgg()}}
\item \href{#method-HydroImport-rollingAggs}{\code{HydroImportFactory$rollingAggs()}}
\item \href{#method-HydroImport-meta}{\code{HydroImportFactory$meta()}}
\item \href{#method-HydroImport-dayStats}{\code{HydroImportFactory$dayStats()}}
\item \href{#method-HydroImport-flowDuration}{\code{HydroImportFactory$flowDuration()}}
\item \href{#method-HydroImport-quality}{\code{HydroImportFactory$quality()}}
\item \href{#method-HydroImport-clone}{\code{HydroImportFactory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-new"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-new}{}}}
\subsection{Method \code{new()}}{
Initialise the new HydroImport object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$new(
  data = NA,
  dataType = "Raw Import",
  modifications = NA,
  stationName = NA,
  riverName = NA,
  WISKI = NA,
  RLOID = NA,
  stationGuide = NA,
  baseURL = NA,
  dataURL = NA,
  measureURL = NA,
  idNRFA = NA,
  urlNRFA = NA,
  easting = NA,
  northing = NA,
  latitude = NA,
  longitude = NA,
  area = NA,
  parameter = NA,
  unitName = NA,
  unit = NA,
  datum = NA,
  boreholeDepth = NA,
  aquifer = NA,
  start = NA,
  end = NA,
  timeStep = NA,
  timeZone = "GMT - UTC±00:00",
  records = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Raw data}

\item{\code{dataType}}{Details the type of data in this environment}

\item{\code{modifications}}{Details modifications made to the data}

\item{\code{stationName}}{Name of gauge}

\item{\code{riverName}}{River name}

\item{\code{WISKI}}{WISKI ID}

\item{\code{RLOID}}{River Levels on the Internet ID}

\item{\code{stationGuide}}{Station Unique Identifier}

\item{\code{baseURL}}{Base URL used in the API}

\item{\code{dataURL}}{End of URL that downloaded the raw data}

\item{\code{measureURL}}{Primarily used in gaining the metadata}

\item{\code{idNRFA}}{National River Flow Archive station identifier}

\item{\code{urlNRFA}}{National River Flow Archive station URL}

\item{\code{easting}}{Easting coordinate}

\item{\code{northing}}{Northing coordinate}

\item{\code{latitude}}{Latitude coordinate}

\item{\code{longitude}}{Longitude coordinate}

\item{\code{area}}{Catchment area of flow/level gauge}

\item{\code{parameter}}{Details the collected data e.g. flow, level etc.}

\item{\code{unitName}}{Unit name used}

\item{\code{unit}}{Details on the unit of measurement}

\item{\code{datum}}{Datum of the gauge, if available}

\item{\code{boreholeDepth}}{Depth of borehole recording}

\item{\code{aquifer}}{Details on the aquifer type}

\item{\code{start}}{Function calculates the date of the first imported record}

\item{\code{end}}{Function calculates the date of the last imported record}

\item{\code{timeStep}}{Function that calculates the timestep in seconds}

\item{\code{timeZone}}{Time zone used in data, defaults to GMT}

\item{\code{records}}{Function calculates the number of records}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-print"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-print}{}}}
\subsection{Method \code{print()}}{
Display the R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$print(.)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-methods"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-methods}{}}}
\subsection{Method \code{methods()}}{
Display the methods available in the R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$methods(.)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-summary"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-summary}{}}}
\subsection{Method \code{summary()}}{
Display a summary of the R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$summary(.)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-asVol"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-asVol}{}}}
\subsection{Method \code{asVol()}}{
Calculate the volume of water for flow or rainfall data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$asVol(.)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-hydroYearDay"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-hydroYearDay}{}}}
\subsection{Method \code{hydroYearDay()}}{
Calculate hydrological year and day
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$hydroYearDay(.)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-rmVol"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-rmVol}{}}}
\subsection{Method \code{rmVol()}}{
Remove the calculated volume
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$rmVol(.)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-rmHY"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-rmHY}{}}}
\subsection{Method \code{rmHY()}}{
Remove the calculated hydroYear
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$rmHY(.)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-rmHYD"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-rmHYD}{}}}
\subsection{Method \code{rmHYD()}}{
Remove the calculated hydrological day
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$rmHYD(.)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-coords"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-coords}{}}}
\subsection{Method \code{coords()}}{
Return the coordinates of the gauge
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$coords(.)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-nrfa"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-nrfa}{}}}
\subsection{Method \code{nrfa()}}{
Return the NRFA details of the gauge
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$nrfa(.)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-dataAgg"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-dataAgg}{}}}
\subsection{Method \code{dataAgg()}}{
Return aggregated data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$dataAgg(type = NULL, method = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Set the aggrgegation level to either hourly', 'daily', 'monthly', 'annual', or 'hydroYear'}

\item{\code{method}}{Available aggregation methods include 'min', 'max', 'mean', 'median', and 'sum'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-rollingAggs"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-rollingAggs}{}}}
\subsection{Method \code{rollingAggs()}}{
Return the different user selecting rolling aggregations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$rollingAggs(
  method = "mean",
  rolling_aggregations = c(1, 2, 3, 4, 8, 24, 120),
  interval = 0.25
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Choose mean, median, min, max, and sum for volumes}

\item{\code{rolling_aggregations}}{The hourly aggregations selected by the user}

\item{\code{interval}}{Corrects for the time step of the data, set to 0.25 hours}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-meta"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-meta}{}}}
\subsection{Method \code{meta()}}{
Returns the metadata as a data.table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$meta(.)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-dayStats"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-dayStats}{}}}
\subsection{Method \code{dayStats()}}{
Daily statistics for imported data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$dayStats(plot = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot}}{Set to TRUE, will produce a plot of the statistics}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-flowDuration"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-flowDuration}{}}}
\subsection{Method \code{flowDuration()}}{
Flow duration curve of data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$flowDuration(perc = c(5, 25, 75, 95))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{perc}}{Determine flow percentiles. Set to c(5, 25, 75, 95)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-quality"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-quality}{}}}
\subsection{Method \code{quality()}}{
Displays the number of observations under each quality flag
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$quality(.)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HydroImport-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HydroImport-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HydroImportFactory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
