% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroLoad.R
\name{hydroLoad}
\alias{hydroLoad}
\title{R6 object for imported data via the EA API}
\value{
A class containing raw data, metadata and methods
}
\description{
This is the base class for all data imported via the API. A hydroLoad contains public raw data and private metadata.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Imported data via the API tool. Uses data.table.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hydroLoad-new}{\code{hydroLoad$new()}}
\item \href{#method-hydroLoad-print}{\code{hydroLoad$print()}}
\item \href{#method-hydroLoad-methods}{\code{hydroLoad$methods()}}
\item \href{#method-hydroLoad-summary}{\code{hydroLoad$summary()}}
\item \href{#method-hydroLoad-asVol}{\code{hydroLoad$asVol()}}
\item \href{#method-hydroLoad-hydroYearDay}{\code{hydroLoad$hydroYearDay()}}
\item \href{#method-hydroLoad-rmVol}{\code{hydroLoad$rmVol()}}
\item \href{#method-hydroLoad-rmHY}{\code{hydroLoad$rmHY()}}
\item \href{#method-hydroLoad-rmHYD}{\code{hydroLoad$rmHYD()}}
\item \href{#method-hydroLoad-coords}{\code{hydroLoad$coords()}}
\item \href{#method-hydroLoad-nrfa}{\code{hydroLoad$nrfa()}}
\item \href{#method-hydroLoad-hourlyAgg}{\code{hydroLoad$hourlyAgg()}}
\item \href{#method-hydroLoad-dailyAgg}{\code{hydroLoad$dailyAgg()}}
\item \href{#method-hydroLoad-monthlyAgg}{\code{hydroLoad$monthlyAgg()}}
\item \href{#method-hydroLoad-annualAgg}{\code{hydroLoad$annualAgg()}}
\item \href{#method-hydroLoad-hydroYearAgg}{\code{hydroLoad$hydroYearAgg()}}
\item \href{#method-hydroLoad-rollingAggs}{\code{hydroLoad$rollingAggs()}}
\item \href{#method-hydroLoad-meta}{\code{hydroLoad$meta()}}
\item \href{#method-hydroLoad-quality}{\code{hydroLoad$quality()}}
\item \href{#method-hydroLoad-clone}{\code{hydroLoad$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-new"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-new}{}}}
\subsection{Method \code{new()}}{
Initialise the new hydroLoad object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$new(
  data = NA,
  stationName = NA,
  riverName = NA,
  WISKI = NA,
  RLOID = NA,
  stationGuide = NA,
  baseURL = NA,
  dataURL = NA,
  measureURL = NA,
  idNRFA = NA,
  urlNRFA = NA,
  easting = NA,
  northing = NA,
  latitude = NA,
  longitude = NA,
  area = NA,
  parameter = NA,
  unitName = NA,
  unit = NA,
  datum = NA,
  boreholeDepth = NA,
  aquifer = NA,
  start = NA,
  end = NA,
  timeZone = "GMT - UTC±00:00",
  records = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Raw data}

\item{\code{stationName}}{Name of gauge}

\item{\code{riverName}}{River name}

\item{\code{WISKI}}{WISKI ID}

\item{\code{RLOID}}{River Levels on the Internet ID}

\item{\code{stationGuide}}{Station Unique IDentifier}

\item{\code{baseURL}}{Base URL used in the API}

\item{\code{dataURL}}{End of URL that downloaded the raw data}

\item{\code{measureURL}}{Primarily used in gaining the metadata}

\item{\code{idNRFA}}{National River Flow Archive station identifier}

\item{\code{urlNRFA}}{National River Flow Archive station URL}

\item{\code{easting}}{Easting coordinate}

\item{\code{northing}}{Northing coordinate}

\item{\code{latitude}}{Latitude coordinate}

\item{\code{longitude}}{Longitude coordinate}

\item{\code{area}}{Catchment area of flow/level gauge}

\item{\code{parameter}}{Details the collected data e.g. flow, level etc.}

\item{\code{unitName}}{Unit name used}

\item{\code{unit}}{Details on the unit of measurement}

\item{\code{datum}}{Datum of the gauge, if available}

\item{\code{boreholeDepth}}{Depth of borehole recording}

\item{\code{aquifer}}{Details on the aquifer type}

\item{\code{start}}{Function calculates the date of the first imported record}

\item{\code{end}}{Function calculates the date of the last imported record}

\item{\code{timeZone}}{Time zone used in data, defaults to GMT}

\item{\code{records}}{Function calculates the number of records}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-print"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-print}{}}}
\subsection{Method \code{print()}}{
Display the R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-methods"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-methods}{}}}
\subsection{Method \code{methods()}}{
Display the methods available in the R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$methods(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-summary"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-summary}{}}}
\subsection{Method \code{summary()}}{
Display a summary of the R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-asVol"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-asVol}{}}}
\subsection{Method \code{asVol()}}{
Calculate the volume of water for flow or rainfall data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$asVol(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-hydroYearDay"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-hydroYearDay}{}}}
\subsection{Method \code{hydroYearDay()}}{
Calculate hydrological year and day
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$hydroYearDay(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-rmVol"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-rmVol}{}}}
\subsection{Method \code{rmVol()}}{
Remove the calculated volume
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$rmVol(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-rmHY"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-rmHY}{}}}
\subsection{Method \code{rmHY()}}{
Remove the calculated hydroYear
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$rmHY(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-rmHYD"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-rmHYD}{}}}
\subsection{Method \code{rmHYD()}}{
Remove the calculated hydrological day
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$rmHYD(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-coords"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-coords}{}}}
\subsection{Method \code{coords()}}{
Return the coordinates of the gauge
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$coords(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-nrfa"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-nrfa}{}}}
\subsection{Method \code{nrfa()}}{
Return the NRFA details of the gauge
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$nrfa(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-hourlyAgg"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-hourlyAgg}{}}}
\subsection{Method \code{hourlyAgg()}}{
Return the hourly aggregated data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$hourlyAgg(method = "mean")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Choose mean, median, min, max, and sum for volumes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-dailyAgg"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-dailyAgg}{}}}
\subsection{Method \code{dailyAgg()}}{
Return the daily aggregated data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$dailyAgg(method = "mean")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Choose mean, median, min, max, and sum for volumes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-monthlyAgg"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-monthlyAgg}{}}}
\subsection{Method \code{monthlyAgg()}}{
Return the monthly aggregated data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$monthlyAgg(method = "mean")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Choose mean, median, min, max, and sum for volumes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-annualAgg"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-annualAgg}{}}}
\subsection{Method \code{annualAgg()}}{
Return the annually aggregated data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$annualAgg(method = "mean")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Choose mean, median, min, max, and sum for volumes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-hydroYearAgg"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-hydroYearAgg}{}}}
\subsection{Method \code{hydroYearAgg()}}{
Return the aggregated data by hydrological year
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$hydroYearAgg(method = "mean")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Choose mean, median, min, max, and sum for volumes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-rollingAggs"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-rollingAggs}{}}}
\subsection{Method \code{rollingAggs()}}{
Return the different user selecting rolling aggregations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$rollingAggs(
  method = "mean",
  rolling_aggregations = c(1, 2, 3, 4, 8, 24, 120),
  interval = 0.25
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Choose mean, median, min, max, and sum for volumes}

\item{\code{rolling_aggregations}}{The hourly aggregations selected by the user}

\item{\code{interval}}{Corrects for the time step of the data, set to 0.25 hours}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-meta"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-meta}{}}}
\subsection{Method \code{meta()}}{
Returns the metadata as a data.table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$meta(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-quality"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-quality}{}}}
\subsection{Method \code{quality()}}{
Displays the number of observations under each quality flag
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$quality(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hydroLoad-clone"></a>}}
\if{latex}{\out{\hypertarget{method-hydroLoad-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hydroLoad$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
