% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroYearDay.R
\name{hydroYearDay}
\alias{hydroYearDay}
\alias{hydroYearDay.data.table}
\alias{hydroYearDay.default}
\title{HydroYearDay}
\usage{
hydroYearDay(x, calendar = "oct_us_gb", ...)

\method{hydroYearDay}{data.table}(x, calendar = "oct_us_gb")

\method{hydroYearDay}{default}(x, hy_cal = "oct_us_gb")
}
\arguments{
\item{hy_cal}{hydrological year calendar. Set to 'oct_us_gb' (USA and UK), but can also be 'sep_br'
(Brazil), 'apr_cl' (Chille).}

\item{d}{Array of dates of class Date}
}
\value{
Creates a hydrological year and hydrological day column
}
\description{
Creates fields for hydrological year and hydrological day.

Speed is slow, needs conversion to C++ in future iterrations.
}
\examples{
hydroYearDay(Buildwas$GaugeData$DateTime, hy_cal = "oct_us_gb")
}
