% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkEvents.R
\name{linkEvents}
\alias{linkEvents}
\title{A means to pair up hydrological events by nearest match time}
\usage{
linkEvents(
  upstream = NULL,
  downstream = NULL,
  timeMin = 0,
  timeMax = 1000,
  matrix = TRUE
)
}
\arguments{
\item{upstream}{Peak values and date time of upstream site}

\item{downstream}{Peak values and date time of downstream site}

\item{timeMin}{Used to filter the time difference in paired peaks, set as 0}

\item{timeMax}{Used to filter the time difference in paired peaks, set as 1000}

\item{matrix}{Set to true, exports a matrix of difftimes in minutes that can
be used to highlight the peak matching. If set as FALSE a nearest join is used.}
}
\value{
a data.table of paired events
}
\description{
A means to pair up hydrological events by nearest match time
}
\examples{
bewdS <- loadAPI(ID = '2001', 'level', 900, 'instantaneous',
                 datapoints = 'range',
                 from = '2008-10-01 09:00')
                 bewdS$hydroYearDay()

buildS <- loadAPI(ID = '2134', 'level', 900, 'instantaneous',
                  datapoints = 'range',
                  from = '2008-10-01 09:00')
buildS$hydroYearDay()

bewdPS <- findPeaks(bewdS, levels = 100)
buildPS <- findPeaks(buildS, levels = 100)
pairedStage <- linkEvents(upstream = buildPS,
                          downstream = bewdPS)
pairedStage

## Plot the paired events
plot(pairedStage$value ~ pairedStage$`i.value`,
     xlab = 'Buildwas Stage (m)',
     ylab = "Bewdley Stage (m)")
model <- lm(pairedStage$value ~ pairedStage$`i.value`)
abline(model, lwd = 2, col = 2)
}
