% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAPIDep.R
\name{loadAPIDep}
\alias{loadAPIDep}
\title{Depreciated version Pull data from the Environment Agency's API}
\usage{
loadAPIDep(
  ID = NULL,
  measure = NULL,
  period = NULL,
  type = NULL,
  datapoints = "standard",
  from = NULL,
  to = NULL,
  lat = NULL,
  long = NULL,
  easting = NULL,
  northing = NULL,
  dist = NULL,
  obsProperty = NULL,
  meta = TRUE,
  rtExt = FALSE,
  rtLookup = FALSE
)
}
\arguments{
\item{ID}{Use to specify a particular WISKI ID or for when downloading all
stations 'wiski' or 'nrfa' to filter results. Additionally to access the
real time APIs use "flood" for all critical sites and "tidal" for the
category A tide gauges.}

\item{measure}{Use this when exporting observations to select the available
parameter. Generally 'flow', 'level', or 'groundwater'}

\item{period}{This is so you can select the time steps available, generally
900 (15 min) or 86400 (daily)}

\item{type}{Selects the data type, instantaneous, min, max etc.}

\item{datapoints}{Allows you to decide where in the time series you are
pulling the data from}

\item{from}{First time step of choice}

\item{to}{Last time step of choice}

\item{lat}{Latitude coordinate for when searching by geographic area}

\item{long}{Longitude coordinate for when searching by geographic area}

\item{easting}{Easting coordinate for when searching by geographic area}

\item{northing}{Northing coordinate for when searching by geographic area}

\item{dist}{Distance in km for how far you wish to search}

\item{obsProperty}{Used to filter the stations when identifying sites.
Available metrics;
"waterFlow", "waterLevel", "rainfall", "groundwaterLevel", "wind",
"temperature", "ammonium", "dissolved-oxygen", "conductivity", "ph",
"temperature", "turbidity", "nitrate", "chlorophyll", "salinity",
"bga", "fdom"}

\item{meta}{Set as TRUE, exports metadata for gauge with the data request}

\item{rtExt}{Set to NULL, if TRUE hydrological data series are extended with
the real time API}

\item{rtLookup}{Set as FALSE, this uses an inbuilt dataset, if set to TRUE
this will update the table using the most recent data. However, this can
dramatically increase the run time.}
}
\value{
Various outputs see vignettes
}
\description{
The loadAPI function is used to interogate the EAs API. Data for
all station can be pulled, or when specified to a single site various
datasets can be pulled.
}
\examples{
#loadAPI()

loadAPI(easting = 378235, northing = 276165, dist = 30, ID = "nrfa")

loadAPI(ID = "L1207")

loadAPI(
  ID = "L1207",
  measure = "level",
  period = 900,
  type = "instantaneous",
  datapoints = "earliest"
)

dt <- loadAPI(
  ID = "L1207",
  measure = "level",
  period = 900,
  type = "instantaneous",
  datapoints = "all"
)

with(dt$data, plot(value ~ dateTime,
  type = "l",
  xlab = "Time",
  ylab = "Stage (mAoD)"
))
}
