% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeData.R
\name{mergeData}
\alias{mergeData}
\title{Merge HydroImport and HydroAggs R6 objects}
\usage{
mergeData(..., metadata = FALSE)
}
\arguments{
\item{...}{Datasets downloaded from the EAs API set inside an R6 container}

\item{metadata}{Set to FALSE, if TRUE collated metadata are exported with the
merged dataset}
}
\value{
A merged dataset as a data,table, if metadata is TRUE then a list of
2 data.tables
}
\description{
The \code{mergeData()} is used to combine multiple R6 encapsulations.
The data have to be in \code{HydroImport} or \code{HydroAggs} classes. Depending on the
metadata argument, metadata can be exported with the merged dataset as a
list. The merging process is carried out on looped full joins, this allows
datasets f varying lengths to be collated. Missing information is padded
with NAs.
}
\examples{
a <- loadAPI(ID = 2001,
             measure = 'flow',
             period = 900,
             type = 'instantaneous',
             datapoints = 'range',
             from = '2023-06-12 09:00',
             to = '2023-06-13 09:00')

b <- loadAPI(ID = 2002,
             measure = 'flow',
             period = 900,
             type = 'instantaneous',
             datapoints = 'range',
             from = '2023-06-12 09:00',
             to = '2023-07-12 09:00')

c <- loadAPI(ID = 2004,
             measure = 'flow',
             period = 900,
             type = 'instantaneous',
             datapoints = 'range',
             from = '2023-06-13 09:00',
             to = '2023-07-12 09:00')

d <- loadAPI(ID = 2001,
             measure = 'level',
             period = 900,
             type = 'instantaneous',
             datapoints = 'range',
             from = '2023-06-12 09:00',
             to = '2023-07-12 09:00')

z <- mergeData(a, b, c, metadata = FALSE)
z
}
