% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyAgg.R
\name{monthlyAgg}
\alias{monthlyAgg}
\alias{monthlyAgg.data.table}
\title{monthlyAgg}
\usage{
monthlyAgg(x, method = "mean", ...)

\method{monthlyAgg}{data.table}(x, method = "mean", ...)
}
\arguments{
\item{x}{Data generated in the riskyData package}

\item{method}{'mean', 'median', 'max', 'min', or 'sum'}

\item{...}{Other variables as required}
}
\value{
An aggregated dataset taken to the calendar month
}
\description{
Aggregates sub monthly time series into a monthly resolution,
aggregations are carried out to the calendar day not a rolling time period.
}
\examples{
monthlyAgg(Buildwas)
}
