% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naRuns.R
\name{naRun}
\alias{naRun}
\title{This identifies where NA values in the time series sit}
\usage{
naRun(dateTime = NULL, value = NULL, timestep = 900)
}
\arguments{
\item{dateTime}{The date and time series}

\item{value}{The observed data}

\item{timestep}{The time step used applied in the observed data in seconds}
}
\value{
A data.table containing the start and end points of NA runs
}
\description{
This identifies where NA values in the time series sit
}
\examples{
data(bewdley)
nas <- naRun(bewdley$data$dateTime,
             bewdley$data$value,
             timestep = 900)
nas
}
