% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollingAggs.R
\name{rollingAggs}
\alias{rollingAggs}
\alias{rollingAggs.data.table}
\title{rollingAggs}
\usage{
rollingAggs(x, rolls = NULL, interval = 0.25, method = "mean")

\method{rollingAggs}{data.table}(x, rolls = NULL, interval = 0.25, method = "mean")
}
\arguments{
\item{rolls}{User defined periods of aggregation}

\item{interval}{Set as 0.25 to represent 15 minute data, for hourly change
to 1 etc.}

\item{method}{'min', 'max', 'mean', 'median', and 'sum' options available}

\item{dt}{flowLoad, rainLoad, or stageLoad data}
}
\value{
A dataset containg numerous aggregations on a roll
}
\description{
Provides a rolling aggregation of a user defined periods.
}
\examples{
rollingAggs(Buildwas)
}
