% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtAPI.R
\name{rtAPI}
\alias{rtAPI}
\title{Access the realtime API}
\usage{
rtAPI(ID = NULL, measure = NULL)
}
\arguments{
\item{ID}{Station Reference ID}

\item{measure}{Type of measurement required}
}
\value{
A data.table of data exported from the EAs realtime API
}
\description{
Access the realtime API
}
\examples{
## Generate a table of all IDs available
stationIDs <- riskyData::loadAPI("flood")

## Find parameters available
rtAPI(ID = "2001")
## Download flow data
data <- rtAPI(ID = "2001", measure = "flow")
plot(data$value, type = 'l')
}
