% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snipRange.R
\name{snipRange}
\alias{snipRange}
\title{snipRange used in the loadAPI function to correct "range" datapoints}
\usage{
snipRange(x = NULL, start = NULL, end = NULL)
}
\arguments{
\item{x}{A data.table used in the loadAPI function}

\item{start}{Start date/time}

\item{end}{End date/time, if kept blank will default to last time step}
}
\value{
A data.table of outputs
}
\description{
Extract windows of data from data.table using start and end date/times
}
\examples{
## CODE NOT RUN
}
