% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asVol.R
\name{asVol}
\alias{asVol}
\alias{asVol.data.table}
\alias{asVol.HydroImport}
\alias{asVol.HydroAggs}
\title{Add convert flow or rainfall data to a volume}
\usage{
asVol(x, area)

\method{asVol}{data.table}(x, area = NULL)

\method{asVol}{HydroImport}(x, area = NULL)

\method{asVol}{HydroAggs}(x, area = NULL)
}
\arguments{
\item{x}{Data set of classes \code{HydroImport} or ¬\code{HydroAggs}}

\item{area_km}{Area of coverage for rain gauges}
}
\value{
Calculated volume of flow or rainfall
}
\description{
Add convert flow or rainfall data to a volume
}
\examples{
# Load bewdley dataset
data(bewdley)

# Calculate volumes
asVol(bewdley)

# Print the bewdley dataset
bewdley
}
