% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchAvg.R
\name{catchAvg}
\alias{catchAvg}
\title{Calculate catchment average rainfall}
\usage{
catchAvg(..., areas = NULL)
}
\arguments{
\item{...}{hydroImport containers}

\item{areas}{user specified area sizes contributed by each rain gauge}
}
\value{
Catchment averaged rainfall in data.table format
}
\description{
Calculate catchment average rainfall
}
\examples{
## Do not run
}
