% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportToPDM.R
\name{exportToPDM}
\alias{exportToPDM}
\title{Export data in a format suitable for PDM for PCs}
\usage{
exportToPDM(x = NULL, path = NULL, type = "csv", name = "PDM_Input")
}
\arguments{
\item{x}{File, currently only supports the class \code{catchAvg} for rainfall}

\item{path}{Optional path for file export. Defaults to working directory.}

\item{type}{Export type that supports csv, XML, and data.table (for internal use)}

\item{name}{Catchment name used in XML exports}
}
\value{
csv, XML or data.tables in the format suitable for PDM for PCs
}
\description{
Export data in a format suitable for PDM for PCs
}
\examples{
## Import some data
boot <- loadAPI(ID = "590310",
                measure = 'rainfall',
                period = 900,
                type = 'total',
                datapoints = 'range',
                from = '2023-10-01 09:00',
                to = '2024-10-01 08:45')

whit <- loadAPI(ID = "197036",
                measure = 'rainfall',
                period = 900,
                type = 'total',
                datapoints = 'range',
                from = '2023-10-01 09:00',
                to = '2024-10-01 08:45')

dogd <- loadAPI(ID = "U06",
                measure = 'rainfall',
                period = 900,
                type = 'total',
                datapoints = 'range',
                from = '2023-10-01 09:00',
                to = '2024-10-01 08:45')
## Carry oout a catchment average
rainfall <- catchAvg(boot, whit, dogd, areas = c(45, 25, 15))

exportToPDM(x = rainfall, type = "data.table")
}
