% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeta.R
\name{getMeta}
\alias{getMeta}
\title{Import metadata for a gauge}
\usage{
getMeta(ID = NULL, mainLink = NULL, measureLink = NULL, import = NULL)
}
\arguments{
\item{ID}{Use to specify a particular WISKI ID}

\item{mainLink}{URL to the main data download}

\item{measureLink}{URL to the measures element of the site}

\item{import}{The main raw data import file, used to collate dates etc.}
}
\value{
A data.table of gauge metadata, for adding to the R6 \code{hydroLoad}
}
\description{
Import metadata for a gauge
}
\examples{
#getMeta(2001)
}
