% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourlyAgg.R
\name{hourlyAgg}
\alias{hourlyAgg}
\alias{hourlyAgg.data.table}
\alias{hourlyAgg.HydroImport}
\title{hourlyAgg}
\usage{
hourlyAgg(x, method = "mean", ...)

\method{hourlyAgg}{data.table}(x, method = "mean", ...)

\method{hourlyAgg}{HydroImport}(x, method = "mean", ...)
}
\arguments{
\item{x}{Data generated in the riskyData package}

\item{method}{'mean', 'median', 'max', 'min', or 'sum'}

\item{...}{Other variables as required}
}
\value{
A dataset of hourly aggregated, data are not stored in a \code{HydroAggs} container.
}
\description{
Aggregates sub hourly time series into an hourly resolution,
aggregations are carried out to the calendar day not a rolling time period.
}
\examples{
data(bewdley)

## Calculate median flow
hourlyAgg(bewdley, method = "median")
}
