% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importHyrad.R
\name{importHYRAD}
\alias{importHYRAD}
\title{Import Hyrad data}
\usage{
importHYRAD(path = NULL, expand = FALSE)
}
\arguments{
\item{path}{File path to desired file}

\item{expand}{Argument to control whether to expand to 15min series or not}
}
\value{
Returns a data.table in the PDM for PCs format
}
\description{
Import Hyrad data
}
\examples{
importHYRAD(files[1], expand = TRUE)
importHYRAD(files[1], expand = FALSE)
}
