% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmFunctions.R
\name{rmHY}
\alias{rmHY}
\alias{rmHY.data.table}
\title{Remove hydrological year from hydroYear class}
\usage{
rmHY(x)

\method{rmHY}{data.table}(x)
}
\arguments{
\item{x}{Raw data in data.table format in hydroLoad class}
}
\description{
Remove hydrological year from hydroYear class
}
\examples{
# obj$rmHY()
}
