% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmFunctions.R
\name{rmHYD}
\alias{rmHYD}
\alias{rmHYD.data.table}
\title{Remove hydrological day of year from hydroYear class}
\usage{
rmHYD(x)

\method{rmHYD}{data.table}(x)
}
\arguments{
\item{x}{Raw data in data.table format in hydroLoad class}
}
\value{
The hydroYearDay column removed from the \code{obj$data}
}
\description{
Remove hydrological day of year from hydroYear class
}
\examples{
 # obj$rmHYD()
}
