% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmFunctions.R
\name{rmVol}
\alias{rmVol}
\title{Remove calculated volume from hydroYear class}
\usage{
rmVol(x)
}
\arguments{
\item{x}{Raw data in data.table format in hydroLoad class}
}
\value{
The volume column removed from the \code{obj$data}
}
\description{
Remove calculated volume from hydroYear class
}
\examples{
# obj$rmVol()
}
